# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Company API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  ok: boolean;
  timestamp: string;
  database: string;
}

export interface OpenRouterTestResult {
  success: boolean;
  model?: string;
  text?: string;
  error?: string;
}

export interface CreateRunInput {
  userKeywords?: string;
}

export interface Run {
  id: string;
  createdAt?: string;
  updatedAt?: string;
  status: string;
  phase: string;
  userKeywords?: string;
  companyName?: string;
  companyTagline?: string;
  selectedIdeaTitle?: string;
  selectedIdeaSummary?: string;
  winnerAgentKey?: string;
}

export interface TranscriptMessage {
  id: string;
  runId: string;
  createdAt?: string;
  phase?: string;
  agentKey?: string;
  roleType?: string;
  messageType?: string;
  content: string;
  sortOrder: number;
}

export interface Artifact {
  id: string;
  runId: string;
  createdAt?: string;
  artifactType: string;
  title?: string;
  contentText?: string;
  storagePath?: string;
}

export interface PhaseResponse {
  status: string;
  message: string;
  runId?: string;
}

export interface PreviewResponse {
  status: string;
  message?: string;
  html?: string;
}

export interface AgentInfo {
  key: string;
  displayName: string;
  roleType: string;
  defaultModel?: string;
  color: string;
  description?: string;
  avatar?: string;
}

export interface AgentConfig {
  founders: AgentInfo[];
  workers: AgentInfo[];
}