# lib/api-zod/src/generated/types/agentConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Company API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentInfo } from "./agentInfo";

export interface AgentConfig {
  founders: AgentInfo[];
  workers: AgentInfo[];
}