# lib/api-zod/src/generated/types/openRouterTestResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Company API
 * OpenAPI spec version: 0.1.0
 */

export interface OpenRouterTestResult {
  success: boolean;
  model?: string;
  text?: string;
  error?: string;
}