# lib/api-zod/src/generated/types/transcriptMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Company API
 * OpenAPI spec version: 0.1.0
 */

export interface TranscriptMessage {
  id: string;
  runId: string;
  createdAt?: string;
  phase?: string;
  agentKey?: string;
  roleType?: string;
  messageType?: string;
  content: string;
  sortOrder: number;
}